Misc. clean-up
A variety of changes that aim to improve the library in a general sense.

These changes were all made while troubleshooting a bug that I initially
thought was in this library. Ultimately, the bug was in the theme, not
this library, but the changes made to this plugin are still beneficial.